+++ htdocs/js/pages/Job.class.js
@@ -365,7 +365,7 @@ Page.Job = class Job extends Page.PageUtils {
 		
 		// alerts (hidden unless needed)
 		html += '<div class="box toggle" id="d_job_alerts" style="display:none">';
-			html += '<div class="box_title">';
+			html += '<div class="box_title" style="color:var(--red);">';
 				html += '<i></i><span>Server Alerts</span>';
 			html += '</div>';
 			html += '<div class="box_content table">';
@@ -1048,7 +1048,7 @@ Page.Job = class Job extends Page.PageUtils {
 			if (state && state.started) {
 				var stub = { type: 'controller', node, state };
 				if (state.active) rows.push(stub);
-				else if (state.completed) {
+				else if (state.completed || state.error) {
 					stub.completed = state.started; // for table sorting
 					completed_stubs.push(stub);
 				}