+++ htdocs/js/pages/Servers.class.js
@@ -203,7 +203,7 @@ Page.Servers = class Servers extends Page.ServerUtils {
 		html += '</div>'; // box_content
 		
 		html += '<div class="box_buttons">';
-			html += '<div class="button secondary" onClick="$P().go_server_search()"><i class="mdi mdi-magnify">&nbsp;</i>Search History...</div>';
+			html += '<div class="button secondary" onClick="$P().go_server_search()"><i class="mdi mdi-magnify">&nbsp;</i>Server History...</div>';
 			if (app.hasPrivilege('add_servers')) html += '<div class="button default" onClick="$P().showAddServerDialog()"><i class="mdi mdi-plus-circle-outline">&nbsp;</i>Add Server...</div>';
 		html += '</div>'; // box_buttons